seed models

+++ db/migrate/20171106103618_create_stocks.rb
@@ -2,7 +2,7 @@ class CreateStocks < ActiveRecord::Migration[5.1]
   def change
     create_table :stocks do |t|
       t.string :name_item
-      t.integer :barcode
+      t.bigint :barcode
       t.decimal :price
       t.integer :quantity
       t.references :category, foreign_key: true
